pyproject.toml
[project]
name = "food-order"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk",
    "pydantic",
    "python-dotenv",
    "google-genai",
    "litellm",
    "pillow",
    "pytesseract"
]
